+++ R/Spectra.R
@@ -414,7 +414,11 @@ NULL
 #'   is *simplified* to a `numeric` if length of `x` or `y` is one.
 #'
 #' - `filterIntensity`: filters each spectrum keeping only peaks with
-#'   intensities that are within the provided range (parameter `intensity`). To
+#'   intensities that are within the provided range or match the criteria of
+#'   the provided function. For the former, parameter `intensity` has to be a
+#'   `numeric` defining the intensity range, for the latter a `function` that
+#'   takes the intensity values of the spectrum and returns a `logical` whether
+#'   the peak should be retained or not (see examples below for details). To
 #'   remove only peaks with intensities below a certain threshold, say 100, use
 #'   `intensity = c(100, Inf)`. Note: also a single value can be passed with
 #'   the `intensity` parameter in which case an upper limit of `Inf` is used.
@@ -548,8 +552,10 @@ NULL
 #'
 #' @param intensity For `filterIntensity`: `numeric` of length 1 or 2 defining
 #'     either the lower or the lower and upper intensity limit for the
-#'     filtering. Defaults to `intensity = c(0, Inf)` thus only peaks with `NA`
-#'     intensity are removed.
+#'     filtering, or a `function` that takes the intensities as input and
+#'     returns a `logical` (same length then peaks in the spectrum) whether the
+#'     peak should be retained or not. Defaults to `intensity = c(0, Inf)` thus
+#'     only peaks with `NA` intensity are removed.
 #'
 #' @param k For `pickPeaks`: `integer(1)`, number of values left and right of
 #'  the peak that should be considered in the weighted mean calculation.
@@ -841,6 +847,17 @@ NULL
 #' lengths(mz(res))
 #' lengths(mz(data))
 #'
+#' ## In addition it is possible to pass a function to `filterIntensity`: in
+#' ## the example below we want to keep only peaks that have an intensity which
+#' ## is larger than one third of the maximal peak intensity in that spectrum.
+#' keep_peaks <- function(x) {
+#'     x > max(x, na.rm = TRUE) / 3
+#' }
+#' res2 <- filterIntensity(data, intensity = keep_peaks)
+#' intensity(res2)[[1L]]
+#' intensity(data)[[1L]]
+#'
+#'
 #' ## Since data manipulation operations are by default not directly applied to
 #' ## the data but only added to the internal lazy evaluation queue, it is also
 #' ## possible to remove these data manipulations with the `reset` function:
@@ -1054,7 +1071,8 @@ setMethod("setBackend", c("Spectra", "MsBackend"),
               bknds <- bplapply(split(object@backend, f = f), function(z, ...) {
                   backendInitialize(backend,
                                     data = spectraData(z),
-                                    ...)
+                                    ...,
+                                    BPPARAM = SerialParam())
               }, ..., BPPARAM = BPPARAM)
               bknds <- backendMerge(bknds)
               ## That below ensures the backend is returned in its original
@@ -1518,20 +1536,34 @@ setMethod("filterIntensity", "Spectra",
                    msLevel. = unique(msLevel(object))) {
               if (!.check_ms_level(object, msLevel.))
                   return(object)
-              if (length(intensity) == 1)
-                  intensity <- c(intensity, Inf)
-              if (length(intensity) != 2)
-                  stop("'intensity' should be of length specifying a lower ",
-                       "intensity limit or of length two defining a lower and",
-                       " upper limit.")
-              object <- addProcessing(object, .peaks_filter_intensity,
-                                      intensity = intensity,
-                                      msLevel = msLevel.)
-              object@processing <- .logging(
-                  object@processing, "Remove peaks with intensities outside [",
-                  intensity[1], ", ", intensity[2],
-                  "] in spectra of MS level(s) ",
-                  paste0(msLevel., collapse = ", "), ".")
+              if (is.numeric(intensity)) {
+                  if (length(intensity) == 1)
+                      intensity <- c(intensity, Inf)
+                  if (length(intensity) != 2)
+                      stop("'intensity' should be of length specifying a ",
+                           "lower intensity limit or of length two defining ",
+                           "a lower and upper limit.")
+                  object <- addProcessing(object, .peaks_filter_intensity,
+                                          intensity = intensity,
+                                          msLevel = msLevel.)
+                  object@processing <- .logging(
+                      object@processing, "Remove peaks with intensities ",
+                      "outside [", intensity[1], ", ", intensity[2],
+                      "] in spectra of MS level(s) ",
+                      paste0(msLevel., collapse = ", "), ".")
+              } else {
+                  if (is.function(intensity)) {
+                      object <- addProcessing(
+                          object, .peaks_filter_intensity_function,
+                          intensity = intensity, msLevel = msLevel.)
+                      object@processing <- .logging(
+                          object@processing, "Remove peaks based on their ",
+                          "intensities and a user-provided function ",
+                          "in spectra of MS level(s) ",
+                          paste0(msLevel., collapse = ", "), ".")
+                  }
+                  else stop("'intensity' has to be numeric or a function")
+              }
               object
           })